# pkg/columns: improve performance of sort & filter